Add files via upload
BankFind Suite's aggregate historical bank data is the annual summary of financial and structure data for all FDIC-insured institutions. The data can be used to identify and analyze long term trends and to develop benchmarks to evaluate the current conditions of the banking and thrift industries. This annual data is available as of December 31 for the years 1934 to 2019.
MERGERS = Failures: Assisted Merger
REI_CO = Conversions
UNASSIST = Unassisted Mergers
New_Char = New Charters